+++ api/__init__.py
@@ -1,3 +1,2 @@
-
 from .app import app
-from .commands import tree
+from .tree import tree