+++ app/app.py
@@ -19,8 +19,7 @@ from g4f.client import Client as GPTClient
 # Local imports
 from audio.input import STT
 from utils import *
-from tree import Tree
-from data.constants import CONFIG_FILE, PROJECT_FOLDER, PLUGINS_FOLDER
+from data.constants import CONFIG_FILE, PROJECT_FOLDER, PLUGINS_FOLDER, COMMANDS_FILE
 
 # disable playsound logger
 logging.getLogger("playsound").setLevel(logging.ERROR)
@@ -34,8 +33,17 @@ class App:
     For further reference, VA = Voice Assistant
     """
 
-    def __init__(self, api):
+    def __init__(self, api, tree):
         self.api = api
+        self.tree = tree
+
+    def play_startup(self):
+        if self.config["start-up"]["sound-enable"]:
+            playsound(self.config["start-up"]["sound-path"], block=False)
+            log.info("Played startup sound")
+        if self.config["start-up"]["voice-enable"]:
+            self.api.say(parse_config_answers(self.config[f"start-up"]["answers"]))
+            log.info("Played startup voice synthesis")
 
     @staticmethod
     def decorator(func):
@@ -51,16 +59,11 @@ class App:
             log.info("Configuration file loaded")
 
             # play sounds
-            if self.config["startup"]["sound-enable"]:
-                playsound(self.config["startup"]["sound-path"], block=False)
-                log.info("Played startup sound")
-            if self.config["startup"]["voice-enable"]:
-                self.api.say(parse_config_answers(self.config[f"startup"]["answers"]))
-                log.info("Played startup voice synthesis")
+            self.play_startup()
 
             # <!--------------- cleanup directories ---------------!>
 
-            cleanup(f"{PROJECT_FOLDER}/data/music", 10)
+            # cleanup(f"{PROJECT_FOLDER}/data/music", 10)
 
             # <!--------------- pre-init: end ---------------!>
 
@@ -78,10 +81,9 @@ class App:
     def start(self, start_time):
         log.debug("App initialization started")
 
-        self.trigger_timed_needed = self.config["trigger"]["trigger-mode"] != "disabled"
+        self.trigger_timed_needed = self.config["settings"]["trigger"]["trigger-mode"] != "disabled"
 
         # data tree initializing
-        self.tree = Tree()
         self.tree_init()
 
         log.info("Data tree initialized")
@@ -90,17 +92,15 @@ class App:
         self.history = []
         self.request_monitor = MonitoredVariable()
 
-        if not self.config["text-mode"]:
+        if not self.config["settings"]["text-mode"]:
             # voice recognition
             self.stt = STT(self.api.lang)
 
             # restricting recognition by adding grammar made of commands
-            if self.config["speech"]["speech-mode-restricted"]:
+            if self.config["audio"]["stt"]["speech-mode-restricted"]:
                 self.grammar_recognition_restricted_create()
                 self.stt.recognizer = self.stt.set_grammar(f"{PROJECT_FOLDER}/data/grammar/grammar-{self.lang}.txt",
                                                            self.stt.create_new_recognizer())
-
-            self.stt.stream.stop_stream()
 
             log.debug("Speech to text instance initialized")
 
@@ -114,70 +114,76 @@ class App:
         log.debug(f"Start up time: {time.time() - start_time:.6f}")
 
     def run(self):
-        if not self.config["text-mode"]:
+        if not self.config["settings"]["text-mode"]:
             self.recognition_thread = threading.Thread(target=self.recognition)
             # starting the voice recognition
             self.recognition_thread.start()
 
             log.debug(f"Recognition thread started with name: {self.recognition_thread.daemon}")
+        else:
+            while True:
+                self.process_trigger_no_voice(input("Input: "))
 
-    def handle(self, request):
+    def handle(self, request, trigger=None):
+        # TODO make detailed comments
         tracker.reset()
-        self.request_monitor.value = request
+        if (not request or not self.remove_trigger_word(request)) and self.config["settings"]["trigger"]["trigger-mode"] != "disabled":
+            self.request_monitor.value = trigger
 
-        if (not request or not self.remove_trigger_word(request)) and self.config["trigger"]["trigger-mode"] != "disabled":
-            # if the request does not contain anything and the trigger word is required,
-            # it means that the user called for VA with a trigger word
-            # (that got removed by self.remove_trigger_word)
-            # and did not specify the command; therefore, we answer with a default phrase
             self.api.say(parse_config_answers(self.config[f"answers"]["default"]))
         else:
-            # update the history of requests to base some commands on the previous requests
             self.history_update(request)
-            # checking whether the request contains multiple commands
+            self.request_monitor.value = request
 
-            # using processor and tracking time of it's execution
             total, exec_time = track_time(lambda: self.api.__command_processor__(request))
 
-            log.info(f"Commands search time: {exec_time:.6f}")
-            log.debug(f"Total commands: {total}")
+            log.info(f"Commands subdivision time: {exec_time:.6f}")
+            log.debug(f"Total commands: {total}")  # TODO Revise
 
             if len(total) == 1:
-                # if there is only one command, process it
-                self.find_command(total[0], request)
+                result, exec_time = track_time(lambda: self.find_command(total[0], request))
+                log.info(f"Command tree search time: {exec_time:.6f}")
+                if result:
+                    result, exec_time = track_time(lambda: self.execute_commands([result]))
+                    log.info(f"Command execution time: {exec_time:.6f}")
+                else:
+                    self.api.__no_command_callback__(request)
             elif len(total) > 1:
-                # if there are multiple commands, we can't play multiple audios at once as they would overlap.
-                # so, we just play a confirmative phrase, like "Doing that now, sir" or else.
-                if all(tuple(command) not in self.tree.api.__inside_tts_list__ for command in total):
-                    # if one of the commands itself produces some sound,
-                    # then it could interfere with the confirmation phrase,
-                    # so we check for this flag "inside_tts" to decide whether
-                    # it would interfere and whether the multi confirmation sound should be played
-                    self.api.say(parse_config_answers(self.config[f"answers"]["multi"]))
-                # processing each command separately
+                commands_to_execute = []
                 for command in total:
-                    self.find_command(command, request, multi=True)
+                    commands_to_execute.append(self.find_command(command, request))
+
+                amount_of_command_actions = len([x for x in commands_to_execute if x is not None])
+                if amount_of_command_actions > 1:
+                    if not issubset(total, self.tree.__inside_tts_list__):
+                        self.api.say(parse_config_answers(self.config[f"answers"]["multi"]))
+
+                    self.execute_commands(commands_to_execute, multi=True)
+                elif amount_of_command_actions == 1:
+                    self.execute_commands(commands_to_execute)
+                else:
+                    self.api.__no_command_callback__(request)
+
             elif not total:
-                # If something was said by user after the trigger word, but no commands were recognized,
-                # then this phrase is being sent to plugins command callbacks or answering
                 self.api.__no_command_callback__(request)
 
-    def find_command(self, command, full, multi: bool = False):
-        result = self.tree.api.find_command(command)
+    def find_command(self, command, full):
+        result = self.tree.find_command(command)
         log.info(f"Execution parameters: {result}")  # TODO fix that log
 
         if result and not all(element is None for element in result):  # if the command was found, process it
             result = list(result)
             result.extend([command, full])  # a list can be extended
-            if result[2] and not multi:
-                # if the command specifies an answer and only one command is being processed, use the TTS engine.
-                self.api.say(parse_config_answers(result[2]))
-            else:
-                tracker.set_value("✅ Done")
 
-            self.do(result)
+            return result
         else:
-            self.api.__no_command_callback__(full)
+            return None
+
+    def execute_commands(self, commands, multi: bool = False):
+        for command in commands:
+            if command[2] and not multi:
+                self.api.say(parse_config_answers(command[2]))
+            self.do(command)
 
     def recognition(self):
         """
@@ -186,25 +192,35 @@ class App:
         while self.running:
             if self.stt.stream.is_active():
                 for word in self.stt.listen():
-                    if self.trigger_timed_needed:
-                        result = self.remove_trigger_word(word)
-                        if result != "blank":
-                            if self.config["trigger"]["trigger-mode"] == "timed":
-                                self.trigger_timed_needed = False
-                                self.trigger_counter(self.config["trigger"]["trigger-time"])
-                            self.handle(result)
-                    else:
-                        self.handle(word)
+                    self.process_trigger(word)
+
+    def process_trigger(self, request):
+        if self.trigger_timed_needed:
+            trigger, result = self.remove_trigger_word(request)
+            if result != "blank":
+                if self.config["settings"]["trigger"]["trigger-mode"] == "timed":
+                    self.trigger_timed_needed = False
+                    self.trigger_counter(self.config["trigger"]["trigger-time"])
+                self.handle(result, trigger)
+        else:
+            self.handle(word)
+
+    def process_trigger_no_voice(self, request):
+        trigger, result = self.remove_trigger_word(request)
+        if result != "blank":
+            self.handle(result, trigger)
+        else:
+            self.handle(request)
 
     def remove_trigger_word(self, request):
         """
         Removes trigger words from the input
         """
-        for trigger in self.config["trigger"][f"triggers"]:
+        for trigger in self.config["settings"]["trigger"][f"triggers"]:
             if trigger in request:
                 request = " ".join(request.split(trigger)[1:])[1:]
-                return request
-        return "blank"
+                return trigger, request
+        return "blank", "blank"
 
     def trigger_counter(self, times):
         trigger_word_countdown_thread = threading.Timer(times, self.trigger_timed_needed)
@@ -216,8 +232,10 @@ class App:
         log.info("Trigger countdown ended")
 
     def tree_init(self):
-        commands = self.config["commands"]
-        commands_repeat = self.config["commands-repeat"]
+        data = filter_lang_config(self.config["commands"], self.lang)
+
+        commands = data["default"]
+        commands_repeat = data["repeat"]
 
         for command in commands:
             equiv = command.get(f'equivalents', [])
@@ -244,10 +262,10 @@ class App:
                     repeat.get(f"synonyms"),
                 )
 
-    def add_command(self, com: tuple, action: str, parameters: dict = None, synthesize: list = None,
+    def add_command(self, com: tuple, action: str, parameters: dict = None, responses: list = None,
                     synonyms: dict = None, equivalents: list = None, inside_tts: bool = False):
-        self.tree.api.add_commands(
-            {com: {"action": action, "parameters": parameters, "synthesize": synthesize, "synonyms": synonyms,
+        self.tree.add_commands(
+            {com: {"action": action, "parameters": parameters, "responses": responses, "synonyms": synonyms,
                    "equivalents": equivalents, "inside_tts": inside_tts}})
 
     def history_update(self, request):
@@ -257,7 +275,7 @@ class App:
         timestamp = datetime.now().isoformat()  # Get a timestamp
         new_event = {"timestamp": timestamp, "request": request}
 
-        if len(self.history) > self.config.get("max-history-length"):
+        if len(self.history) > self.config["settings"]["max-history-length"]:
             self.history.pop(0)
 
         self.history.append(new_event)
@@ -288,9 +306,9 @@ class App:
         This file is used for a vosk speech-to-text model to speed up the recognition speed and quality
         """
         with open(f"{PROJECT_FOLDER}/data/grammar/grammar-{self.lang}.txt", "w") as file:
-            file.write('["' + " ".join(self.config['trigger'].get(f"triggers")))
-            file.write(self.config["speech"].get(f"restricted-add-line"))
-            file.write(self.tree.api.recognizer_string + '"]')
+            file.write('["' + " ".join(self.config["settings"]["trigger"].get(f"triggers")))
+            file.write(self.config["audio"]["stt"].get(f"restricted-add-line"))
+            file.write(self.tree.recognizer_string + '"]')
 
     # below methods are actions that need access to the main app instance
     # <!--------------------------------------------------------------------!>